pyproject.toml
[project]
name = "speech-analysis-module"
version = "0.0.4"
authors = [
  { name="Emma Rico Martín", email="emmaricomartin@gmail.com" },
]
description = "SAM is a Python library for sound/audio processing in experimental psychology"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "librosa",
    "numpy",
    "scipy",
    "matplotlib",
    "soundfile",
    "sounddevice",
    "pydub",
    "praat-parselmouth",
    "openai-whisper",
    "imageio-ffmpeg",
	"serial",
]
license = "GPL-3.0-only"
license-files = ["license.txt"]

[project.urls]
Homepage =  "https://github.com/emmaricomartin/Speech-Analysis-Module"
#Homepage = "https://github.com/pypa/sampleproject"
#Issues = "https://github.com/pypa/sampleproject/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["speech_analysis_module*"]
